ID3 Parsing: Only convert unicode if framesize > 1
The current id3 parsing code leads to a lock-up of the cpu if the
framesize is 1 and unicode conversion needs to be done. This is since

i = framesize -1

and the conversion loop condition is m<i-1. For framesize = 1 -> i=0 and
the condition m<-1 for unit16_t m keeps the cpu busy for a while and
reads way above the memory boundary of the array 'value' inside the loop.

This patch fixes the problem by not entering the unicode conversion if
the framesize is <= 1.