AudioBuffer: allow setting size of RAM and PSRAM buffer
instead of the fixed buffer size of 300000 for PSRAM and 8000 for
non-PSRAM setup, allow to configure the buffer sizes, example code
for that is
    audio.setBuffersize(32768, 0);
which will set the RAM buffer size to 32k and disable PSRAM usage.
I am investigating "bad WiFi throughput with PSRAM" issues and disabling
PSRAM is one workaround I found.